server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-ai-translation-document==1.1.0b1
azure-cognitiveservices-speech==1.42.0
azure-core==1.32.0
azure-cosmos==4.9.0
azure-identity==1.19.0
azure-storage-blob==12.24.1
beautifulsoup4==4.13.3
black==24.10.0
blinker==1.9.0
bs4==0.0.2
cachelib==0.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
cwcwidth==0.1.9
distlib==0.3.9
dnspython==2.7.0
elevenlabs==1.52.0
et_xmlfile==2.0.0
fastapi==0.115.8
ffmpeg==1.4
filelock==3.17.0
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Session==0.5.0
fsspec==2025.2.0
gcloud==0.18.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
identity==0.5.2
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jwcrypto==1.5.6
llvmlite==0.44.0
lxml==5.3.0
MarkupSafe==3.0.2
mongoengine==0.29.1
more-itertools==10.6.0
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msgspec==0.19.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
oauth2client==4.1.3
openpyxl==3.1.5
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
portalocker==2.10.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.1
pyparsing==3.2.1
Pyrebase4==4.8.0
python-dotenv==0.21.1
python-jwt==4.1.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
semantic-version==2.10.0
setuptools==75.8.2
setuptools-rust==1.10.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
translate-toolkit==3.14.7
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
virtualenv==20.29.1
virtualenvwrapper-win==1.2.7
websockets==15.0
Werkzeug==3.1.3
